fix : api 함수 파일 분배
